Truncation error on POST body; ClusterId resulting in BadRequest. Version 0.2.0.rc0
Truncation error on POST body - ByteBuffer was being reallocated under the
feet of InfluxReporter, and was also being retained between calls for no reason.
Can be simplified by just using a StringBuilder (the body is text+newline)
reallocated on each invocation. Removed ByteBufferOutputStream

ClusterId results in BadRequest - special cased this as "name"

Added 'omit' configuration, and ability to specify no retention
policy (which is then not sent as query param)